Calorie_Tracker/backend/image_recognition.py
"""Image recognition module for calorie extraction."""
from abc import ABC, abstractmethod
from typing import Optional
from PIL import Image
import io

from ..domain import ImageRecognitionResult, FoodItemDetection


class ImageRecognizer(ABC):
    """Abstract base class for image recognition strategies."""
    
    @abstractmethod
    def recognize(self, image_bytes: bytes) -> ImageRecognitionResult:
        """Recognize food in image and return result."""
        pass


class LabelRecognizer(ImageRecognizer):
    """Recognizes nutritional labels in images."""
    
    def recognize(self, image_bytes: bytes) -> ImageRecognitionResult:
        """
        Attempt to extract calorie information from a nutritional label.
        
        Args:
            image_bytes: Raw image bytes
            
        Returns:
            ImageRecognitionResult with extracted calories
        """
        try:
            # TODO: Implement OCR to extract calorie info from label
            # Could use pytesseract, EasyOCR, or cloud API like Google Vision
            
            result = ImageRecognitionResult(
                success=False,
                method="label_recognition",
                error_message="Label recognition not yet implemented"
            )
            return result
        except Exception as e:
            return ImageRecognitionResult(
                success=False,
                method="label_recognition",
                error_message=str(e)
            )


class VisualEstimator(ImageRecognizer):
    """Estimates calories based on visual food detection and generic formulas."""
    
    def recognize(self, image_bytes: bytes) -> ImageRecognitionResult:
        """
        Estimate calories based on detected food items.
        
        Args:
            image_bytes: Raw image bytes
            
        Returns:
            ImageRecognitionResult with estimated calories
        """
        try:
            # TODO: Implement food detection using:
            # - YOLO, Faster R-CNN, or similar object detection model
            # - Could use food-specific models or Google Vision API
            
            result = ImageRecognitionResult(
                success=False,
                method="visual_estimation",
                error_message="Visual estimation not yet implemented"
            )
            return result
        except Exception as e:
            return ImageRecognitionResult(
                success=False,
                method="visual_estimation",
                error_message=str(e)
            )


class ImageProcessor:
    """Main processor for image-based calorie extraction."""
    
    def __init__(self):
        self.label_recognizer = LabelRecognizer()
        self.visual_estimator = VisualEstimator()
    
    def process_image(
        self,
        image_bytes: bytes,
        prefer_method: str = None
    ) -> ImageRecognitionResult:
        """
        Process image to extract or estimate calories.
        
        Args:
            image_bytes: Raw image bytes
            prefer_method: Preferred method ("label" or "visual"), tries preferred first
            
        Returns:
            ImageRecognitionResult
        """
        # Try preferred method first if specified
        if prefer_method == "label":
            result = self.label_recognizer.recognize(image_bytes)
            if result.success:
                return result
            # Fall back to visual estimation
            return self.visual_estimator.recognize(image_bytes)
        
        elif prefer_method == "visual":
            result = self.visual_estimator.recognize(image_bytes)
            if result.success:
                return result
            # Fall back to label recognition
            return self.label_recognizer.recognize(image_bytes)
        
        # Try both methods and return best result
        label_result = self.label_recognizer.recognize(image_bytes)
        visual_result = self.visual_estimator.recognize(image_bytes)
        
        # Prefer label recognition if successful
        if label_result.success:
            return label_result
        return visual_result
    
    def validate_image(self, image_bytes: bytes) -> bool:
        """Validate that bytes contain a valid image."""
        try:
            image = Image.open(io.BytesIO(image_bytes))
            image.verify()
            return True
        except Exception:
            return False
