requirements.txt
streamlit==1.54.0
Pillow==12.1.1
